requirements.txt
langchain
langchain-google-genai
langchain-community
chromadb
pypdf
tiktoken
python-dotenv
